% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_vocabulary}
\alias{transcribeservice_delete_vocabulary}
\title{Deletes a vocabulary from Amazon Transcribe}
\usage{
transcribeservice_delete_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary to delete.}
}
\description{
Deletes a vocabulary from Amazon Transcribe.
}
\section{Request syntax}{
\preformatted{transcribeservice$delete_vocabulary(
  VocabularyName = "string"
)
}
}

\keyword{internal}
